put NWM File client in its own module
